Keep bridge and delete "nothing" platform
I had written the `bridge` platform, intended for use on the Pine chat
network, and the `nothing` platform, which was just for temporary
testing. Because I was done with the `nothing` platform and didn't see a
future use for it, I decided to delete it and not commit it to Git.

However, I mistyped the `rm` command; I'm not sure what exactly
happened, but my guess is that I meant to type `rm platforms/cs_platform_n`
then hit Tab, so bash would autocomplete it to
`rm platforms/cs_platform_nothing.py`. It seems that I accidentally
typed `rm platforms/cs_platform_b` instead; when I autocompleted this,
it gave `platforms/cs_platform_bridge.py`. I don't know that I did this,
but it's the only way I can see that this could have happened; it does
seem likely considering that `b` and `n` are adjacent on QWERTY
keyboards.

Thankfully, I had a small data recovery utility that I had previously
written, and I was able to find a non-erased copy of the code in my
filesystem. It would have been much better, though, to not need to
recover the file, or at least to have had a backup.

The moral of the story is:

Double-check your commands, and make frequent backups!